pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sgcmc_coexistence"
version = "1.0.0"
authors = [
  { name="Keyur Maniya" },
]
description = "A workflow for tracing solid-liquid phase coexistence lines using SGCMC and calphy"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "pyyaml",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
